initial commit
hit a milestone in early dev with implementation for the random access value_iterator deduction and decided to start the repo.
this commit has a mostly-implemented value_iterator class, as well as a couple other miscellaneous iterator types for function mapping.
it also includes the iterator_range class, which has built-in functions for range mapping and accumulation (via std::accumulate).
the test file has test cases for everything currently included, but not every possible combination of primitives (infeasible).