bug fixes + bidirection value iterator
fixed a couple of bugs regarding logical object lifetimes for all the iterator type.

fixed a couple of problems in the function iterator where we assumed V was default constructible.
the V object is now stored as a buffer object, which we manually construct and destroy as needed, thus bypassing the problem.

when i made the mapping iterator i decided the function might be expensive, so i should cache the value.
however that thinking was based on value iterators (since i had just got done making those).
logically, a change to the pointed-to object of the stored iterator of a mapping iterator should be reflected in the next dereference.
thus mapping iterators no longer cache their values.
this also fixes the problem where creating a mapping iterator for an end iterator would dereference it - big no-no i hadn't even thought of until just now.

this has the unfortunate side-effect that dereferencing the mapping iterator is potentially a prvalue, so we can't take its address (and even if we could it'd be rather dangerous).
this means the arrow operator is now out of commission unless i think up a fix.
at the user level, this could be fixed by writing (*a).b instead of a->b, but there are template problems to think of, in addition to lifetime concerns.
the current fix i'm thinking of, but hesitate to implement, is to make the dereference operator store the mapped result to a member variable and return a reference to it.
this would fix the problem but takes up more space in the already-potentially-non-trivial iterator type and could result in a lot of extra calls to the mapping function - though i could just mention that users should avoid doing that.