bug fixes + mapping -> operator
fixed a few lifetime-related bugs - or rather made them less likely to happen.
as it was, you could use operator -> on e.g. an rvalue value_iterator.
the returned address would thus immediately be dangling as the rvalue iterator's lifetime ends just after the function returns.
thus, i added rvalue overloads for the arrow operators for all data-holding iterators and marked them as deleted functions.
this way it's a compile-time error to use them in such a way, which could greatly reduce errors.

decided on the new data contract for the mapping iterator's specific case of using the arrow operator (which is now re-implemented).
everything is now exactly as it was in the last commit, except the addition of a storage location for the mapping return value.
this storage location is only used by the arrow operator.
the object whose address is returned is guaranteed to stay alive at least until the next operator -> call on that iterator object or the end of the iterator's lifetime.
this gives us a little bit of leeway with how it's implemented (e.g. we don't have to copy it when the iterator's copied).
this definition also ensures we have each and every mapping iterator dereference (including ->) use the most up to date value without caching anything like it used to (which could lead to some confusing results if you weren't aware it was happening).