lots of stuff
added a bunch of wrapper functions in the iterator_range class - pretty self-explanatory.

fixed a bunch of problems caused by lacking iterator traits and certain interfaces - should all be good now.
the biggest thing causing problems was the use of std::advance() and other stdlib utilities that need the traits and the interfaces defined.

fixed a bit of weirdness with the difference type potentially being unsigned for value_iterator, which means it's never negative despite being a random access type, which would be weird for users.
additionally it would result in massive (unsigned) numbers when it underflows, which would seem like a bug to me if i was just using this library and didn't know any better.

fixed another similar weirdness where the count for counting iterators was unsigned, which leads to odd behavior when decrementing below zero, which should be allowed for a random access iterator, which a counting iterator could very well be.

aside from all that nonsense, there was only one thing that actually required any new logic:
never in my entire career of C++ have i actually encountered the no-assignable lambda problem - i don't typically ever need to deep copy them.
however, for certain aspects of function iterators and mapping iterators (any fancy iterator type with a stored function), we need to be able to deep-copy the function itself - it's not sufficient to just use the same function for e.g. unary_function_iterator because then the two iterators would point to different objects after incrementing both if the function has internal data (e.g. lambda capture groups).

this was wreaking all kinds of havoc in the lookup wrapper tests for the new iterator_range class.
thus i added a little hacky class to wrap any function like object and make it assignable (so long as it's constructible).
this is called assignable_func<F>.
if F is copy/move assignable, it just does copy/move assignment.
otherwise, it destroys the object explicitly and copy/move constructs a new one in its place.
i don't believe there's any other way to do it, so there it is.