minor changes + more wrapper funcs
fixed the last dashes of lifetime concerns for caching iterators.
made some minor changes to the const/non-const interface for iterators.

i'd like the iterators to be const dereference, but i've realized e.g. for the func iterators this isn't always possible, as the function might an overloaded operator() that modifies internal data and is thus non-const.

added a couple more wrapper functions - they're quite easy to add but getting very tedious.
they're certainly not a priority - only map() is actually important for usability of the library.